// src/model/queryhotelcomment.ts
export class QueryHotelComment {
    MebName: string;
    CardNo: string;
    ReviewsID: number;
    MebID: number;
    KeyID: number;
    FolioID: number;
    Type: number;
    Comment: string;
    Level: number;
    CreateTime: string;
    State: number;
    Flag: number;
    Score: number;
    ImageCount: number;
    ImgUrl: string;
	CommentImg: string;
	StarNumber: string;
}